+++ source/YuzuModDownloader/classes/ModDownloader.cs
@@ -1,5 +1,4 @@
-﻿using HtmlAgilityPack;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
@@ -8,217 +7,208 @@ using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 using System.Windows.Forms;
-using System.Xml;
 
 namespace YuzuModDownloader
 {
-    public sealed class ModDownloader
+    public class ModDownloader
     {
         private const string Quote = "\"";
-        private readonly string _userDirPath;
-        private readonly string _modDirPath;
         private string _sevenZipExePath;
-        
-        public ModDownloader()
+
+        protected ModDownloader()
         {
-            _userDirPath = Directory.Exists("user") ?
+            UserDirPath = Directory.Exists("user") ?
                 Path.Combine(Path.GetDirectoryName(Application.ExecutablePath), "user") :
                 Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "yuzu");
-            _modDirPath = GetModPath(_userDirPath);
+            ModDirectoryPath = GetModPath(UserDirPath);
             _sevenZipExePath = "";
         }
 
-        public delegate void UpdateProgressDelegate(int progressPercentage, string progressText);
+        protected string UserDirPath { get; private set; } = "";
 
-        public event UpdateProgressDelegate UpdateProgress;
+        protected string ModDirectoryPath { get; private set; } = "";
 
-        public async Task DownloadGameTitleIdDatabaseAsync(string fileName)
+        protected internal bool IsModDataLocationToBeDeleted { get; set; } = false;
+
+        protected internal bool IsDownloadedModArchivesToBeDeleted { get; set; } = true;
+
+        protected internal delegate void UpdateProgressDelegate(int progressPercentage, string progressText);
+
+        protected internal event UpdateProgressDelegate UpdateProgress;
+
+        protected void RaiseUpdateProgressDelegate(int progressPercentage, string progressText)
         {
-            // download latest GameTitleID's & place in same directory as executable
-            using (var wc = new WebClient())
+            UpdateProgress?.Invoke(progressPercentage, progressText);
+        }
+
+        /// <summary>
+        /// Downloads all prequisites for YuzuModDownloader to function.
+        /// </summary>
+        /// <returns></returns>
+        protected async Task DownloadPrerequisitesAsync()
+        {
+            if (!IsSevenZipInstalled()) 
+                await DownloadSevenZip();
+        }
+
+        /// <summary>
+        /// Downloads the specified URL and saves it as the file name passed.
+        /// </summary>
+        /// <param name="url">URL of the file to download</param>
+        /// <param name="fileName">Name of the destination local file</param>
+        /// <returns></returns>
+        protected async Task DownloadGameDatabaseAsync(string url, string fileName)
+        {
+            // download game database & place in same directory as executable
+            using (var client = new WebClient())
             {
-                wc.DownloadFileCompleted += (s, e) => UpdateProgress(100, "Done");
-                wc.DownloadProgressChanged += (s, e) => UpdateProgress(e.ProgressPercentage, "Downloading Latest Game TitleID's ...");
-                await wc.DownloadFileTaskAsync("https://raw.githubusercontent.com/amakvana/YuzuModDownloader/main/assets/GameTitleIDs.xml", fileName);
+                client.DownloadFileCompleted += (s, e) => RaiseUpdateProgressDelegate(100, "Done");
+                client.DownloadProgressChanged += (s, e) => RaiseUpdateProgressDelegate(e.ProgressPercentage, $@"Downloading {fileName} ...");
+                await client.DownloadFileTaskAsync(url, fileName);
             }
         }
 
-        public async Task DownloadPrerequisitesAsync()
+        /// <summary>
+        /// Reads the Games list, downloads Mods per game and extracts them into the Mod Data Location.
+        /// </summary>
+        /// <param name="games">List of Games.</param>
+        /// <returns></returns>
+        protected async Task DownloadModsAsync(List<Game> games)
         {
-            // download 7zip so we can use this to extract mods
-            // faster and more cleaner
-            string prerequisitesLocation = $@"{_userDirPath}\7z";
-            string sevenZLocation = $@"{prerequisitesLocation}\7z.zip";
+            using (var client = new WebClient())
+            {
+                foreach (var game in games)
+                {
+                    // clear mod data location if user has checked the option 
+                    if (IsModDataLocationToBeDeleted)
+                    {
+                        DeleteModDataLocationPath(game.ModDataLocation);
+                    }
 
-            // check if 7z is installed locally 
-            // if so, we use that instead of downloading a copy 
-            // faster and uses less bandwidth 
-            bool locallyInstalled7z = false;
+                    foreach (var url in game.ModDownloadUrls)
+                    {
+                        string fileName = url.AbsoluteUri.Substring(url.AbsoluteUri.LastIndexOf('/') + 1).Trim();
+
+                        client.DownloadFileCompleted += (s, e) =>
+                        {
+                            RaiseUpdateProgressDelegate(0, $"Unpacking {fileName} ...");
+
+                            // unzip downloaded mod 
+                            var psi = new ProcessStartInfo
+                            {
+                                CreateNoWindow = true,
+                                WindowStyle = ProcessWindowStyle.Hidden,
+                                FileName = _sevenZipExePath,
+                                Arguments = $@"x {Quote}{game.ModDataLocation}\{fileName}{Quote} -o{Quote}{game.ModDataLocation}{Quote} -aoa"
+                            };
+                            using (var p = Process.Start(psi))
+                            {
+                                p.WaitForExit();
+                                RaiseUpdateProgressDelegate(100, "Done");
+                            }
+                        };
+                        client.DownloadProgressChanged += (s, e) => RaiseUpdateProgressDelegate(e.ProgressPercentage, $"Downloading {fileName} ...");
+                        await client.DownloadFileTaskAsync(url.AbsoluteUri, $@"{game.ModDataLocation}/{fileName}");
+                    }
+                }
+            }
+
+            if (IsDownloadedModArchivesToBeDeleted)
+            {
+                DeleteDownloadedModArchiveFiles();
+            }
+        }
+
+        /// <summary>
+        /// Checks if 7-Zip is currently installed on this machine.
+        /// </summary>
+        /// <returns>
+        /// <c>true</c> if installed; otherwise <c>false</c>.
+        /// </returns>
+        private bool IsSevenZipInstalled()
+        {
             string programFiles = Environment.ExpandEnvironmentVariables("%ProgramW6432%");
             string programFilesX86 = Environment.ExpandEnvironmentVariables("%ProgramFiles(x86)%");
+
             if (programFiles != null && File.Exists($@"{programFiles}\7-Zip\7z.exe"))
             {
                 _sevenZipExePath = $@"{programFiles}\7-Zip\7z.exe";
-                locallyInstalled7z = true;
+                return true;
             }
             if (programFilesX86 != null && File.Exists($@"{programFilesX86}\7-Zip\7z.exe"))
             {
                 _sevenZipExePath = $@"{programFilesX86}\7-Zip\7z.exe";
-                locallyInstalled7z = true;
+                return true;
             }
-
-            // if 7z is not already installed on the machine, download a copy 
-            if (!locallyInstalled7z)
-            {
-                if (Directory.Exists(prerequisitesLocation)) DirectoryUtilities.DeleteDirectory(prerequisitesLocation, true);
-                Directory.CreateDirectory(prerequisitesLocation);
-                using (var wc = new WebClient())
-                {
-                    wc.DownloadFileCompleted += (s, e) =>
-                    {
-                        UpdateProgress(0, "Unpacking prerequisites ...");
-                        using (var archive = ZipFile.OpenRead(sevenZLocation))
-                        {
-                            foreach (var entry in archive.Entries)
-                            {
-                                entry.ExtractToFile(Path.Combine($"{prerequisitesLocation}", entry.FullName), true);
-                            }
-                        }
-                        _sevenZipExePath = $@"{prerequisitesLocation}\7z.exe";
-                        UpdateProgress(100, "Done");
-                    };
-                    wc.DownloadProgressChanged += (s, e) => UpdateProgress(e.ProgressPercentage, "Downloading Prerequisites ...");
-                    //await wc.DownloadFileTaskAsync("https://my.cloudme.com/v1/ws2/:amakvana/:7z/7z.zip", sevenZLocation);
-                    await wc.DownloadFileTaskAsync("https://github.com/amakvana/YuzuModDownloader/raw/main/assets/7z/22.01/7z.zip", sevenZLocation);
-                }
-            }
+            return false;
         }
 
-        public Dictionary<string, string> ReadGameTitleIdDatabase(string gameTitleIDsXml)
+        /// <summary>
+        /// Downloads 7-Zip and extracts it inside the Yuzu User Directory path.
+        /// </summary>
+        /// <returns></returns>
+        private async Task DownloadSevenZip()
         {
-            // detect yuzu user directory 
-            // loop through {ModDirPath} folder & get title names from title Id's
-            // return list 
-            var d = new Dictionary<string, string>();
+            string prerequisitesLocation = $@"{UserDirPath}\7z";
+            string sevenZLocation = $@"{prerequisitesLocation}\7z.zip";
 
-            //var doc = new XmlDocument();
-            //doc.Load(gameTitleIDsXml);
-            //var nodes = doc.DocumentElement.SelectNodes("//*[local-name()='games']/*[local-name()='game']");
-            //foreach (XmlNode node in nodes)
-            //{
-            //    string titleName = node["title_name"].InnerText.Trim();
-            //    string titleId = node["title_id"].InnerText.Trim();
-            //    if (!string.IsNullOrWhiteSpace(titleId) && Directory.Exists($"{ModDirPath}/{titleId}"))
-            //    {
-            //        d.Add(titleName, titleId);
-            //    }
-            //}
-
-            // using XmlReader for better performance - http://codingrules.blogspot.com/2010/06/c-xmlreader.html
-            using (var reader = XmlReader.Create(gameTitleIDsXml))
+            if (Directory.Exists(prerequisitesLocation))
             {
-                while (reader.Read())
+                DeleteDirectory(prerequisitesLocation, true);
+            }
+            Directory.CreateDirectory(prerequisitesLocation);
+
+            using (var client = new WebClient())
+            {
+                client.DownloadFileCompleted += (s, e) =>
                 {
-                    if (reader.IsStartElement())
+                    UpdateProgress(0, "Unpacking prerequisites ...");
+                    using (var archive = ZipFile.OpenRead(sevenZLocation))
                     {
-                        switch (reader.Name)
+                        foreach (var entry in archive.Entries)
                         {
-                            case "title_name":
-                                string titleName = reader.ReadElementContentAsString();
-                                reader.Read();
-                                string titleId = reader.ReadElementContentAsString();
-                                if (!string.IsNullOrWhiteSpace(titleId) && Directory.Exists($"{_modDirPath}/{titleId}"))
-                                {
-                                    d.Add(titleName, titleId);
-                                }
-                                break;
-                            default: break;     //do nothing 
+                            entry.ExtractToFile(Path.Combine($"{prerequisitesLocation}", entry.FullName), true);
                         }
                     }
-                }
-            }
-            return d;
-        }
-
-        public async Task DownloadTitleModsAsync(string titleName, string titleId, string modWebsiteUrl, bool deleteExistingMods)
-        {
-            // fetch all download links for current game
-            // extract them into appropriate directory
-
-            var web = new HtmlWeb();
-            var htmlDoc = web.Load(modWebsiteUrl);
-            var nodes = htmlDoc.DocumentNode.SelectNodes($@"//h3[contains(., {Quote}{titleName}{Quote})]/following::table[1]//td//a");
-
-            // if true, delete existing mods 
-            if (deleteExistingMods)
-            {
-                DeleteModData($@"{_modDirPath}\{titleId}");
-            }
-
-            // download all mods for game 
-            if (nodes != null)
-            {
-                using (var wc = new WebClient())
-                {
-                    foreach (HtmlNode node in nodes)
-                    {
-                        string modName = node.InnerText;
-                        string modDownloadUrl = node.Attributes["href"].Value.Trim();
-                        string fileName = modDownloadUrl.Split('/').Last().Trim();
-
-                        // if url is gamebanana, scrape page and get actual download
-                        //if (modDownloadUrl.Contains("gamebanana.com"))
-                        //{
-                        //    // modUrl = GetGameBananaDownloadUrl(modUrl, out fileName);
-                        //    // NEED to fix when files are extracted from GameBanana packages
-                        //}
-                        //else if (modDownloadUrl.Contains("bit.ly"))
-                        //{
-                        //    // coming soon ...
-                        //}
-
-                        if (modDownloadUrl.EndsWith(".zip") || modDownloadUrl.EndsWith(".rar") || modDownloadUrl.EndsWith(".7z"))
-                        {
-                            wc.DownloadFileCompleted += (s, e) =>
-                            {
-                                UpdateProgress(0, $"Unpacking {fileName} ...");
-
-                                // unzip downloaded mod 
-                                var psi = new ProcessStartInfo
-                                {
-                                    WindowStyle = ProcessWindowStyle.Hidden,
-                                    //FileName = $@"{UserDirPath}\7z\7z.exe",
-                                    FileName = _sevenZipExePath,
-                                    Arguments = $@"x {Quote}{_modDirPath}\{titleId}\{fileName}{Quote} -o{Quote}{_modDirPath}\{titleId}{Quote} -aoa"
-                                };
-                                using (var p = Process.Start(psi))
-                                {
-                                    UpdateProgress(100, "Done");
-                                }
-                            };
-                            wc.DownloadProgressChanged += (s, e) => UpdateProgress(e.ProgressPercentage, $"Downloading {fileName} ...");
-                            await wc.DownloadFileTaskAsync(modDownloadUrl, $"{_modDirPath}/{titleId}/{fileName}");
-                        }
-                    }
-                }
+                    _sevenZipExePath = $@"{prerequisitesLocation}\7z.exe";
+                    UpdateProgress(100, "Done");
+                };
+                client.DownloadProgressChanged += (s, e) => RaiseUpdateProgressDelegate(e.ProgressPercentage, "Downloading Prerequisites ...");
+                await client.DownloadFileTaskAsync("https://github.com/amakvana/YuzuModDownloader/raw/main/assets/7z/22.01/7z.zip", sevenZLocation);
             }
         }
 
-        public void DeleteDownloadedModArchives()
+        /// <summary>
+        /// Delete the Downloaded Mod Archive Files
+        /// </summary>
+        private void DeleteDownloadedModArchiveFiles()
         {
+            if (string.IsNullOrWhiteSpace(ModDirectoryPath))
+                return;
+
             // loop through each title_id
             // delete archives from within title_id folder
-            foreach (var subDirectory in Directory.GetDirectories(_modDirPath))
+            var directories = Directory.GetDirectories(ModDirectoryPath);
+            foreach (var subDirectory in directories)
             {
-                DirectoryUtilities.DeleteFiles(subDirectory, ".zip", ".rar", ".7z");
+                DeleteFiles(subDirectory, ".zip", ".rar", ".7z");
             }
         }
 
-        private void DeleteModData(string path)
+        /// <summary>
+        /// Deletes all subdirectories and files inside the Mod Data Location path
+        /// </summary>
+        /// <param name="path">The absolute path of Mod Data Location</param>
+        private void DeleteModDataLocationPath(string path)
         {
-            DirectoryUtilities.DeleteDirectory(path, true);
+            DeleteDirectory(path, true);
         }
 
+        /// <summary>
+        /// Extracts the Mod Path from the Yuzu 'User' folder.
+        /// </summary>
+        /// <param name="baseDirPath">Absolute directory path to the Yuzu 'User' folder.</param>
+        /// <returns></returns>
         private string GetModPath(string baseDirPath)
         {
             // read in qt-config.ini 
@@ -230,16 +220,46 @@ namespace YuzuModDownloader
             return modPath;
         }
 
-        //private string GetGameBananaDownloadUrl(string url, out string fileName)
-        //{
-        //    // grabs the download url from the gamebanana mod webpage
-        //    var web = new HtmlWeb();
-        //    var htmlDoc = web.Load(url);
-        //    var node = htmlDoc.DocumentNode.SelectSingleNode("//module[@id=\"FilesModule\"]/div//a");
-        //    string downloadUrl = node.Attributes["href"].Value;
-        //    fileName = htmlDoc.DocumentNode.SelectSingleNode("//module[@id=\"FilesModule\"]/div//code").InnerText.Trim();
-        //    return downloadUrl;
-        //}
+        private void DeleteDirectory(string path, bool recursive = false)
+        {
+            // https://stackoverflow.com/a/1288747
+            // Modified to optionally allow recusive deletion
 
+            var dir = new DirectoryInfo(path);
+            var files = dir.GetFiles();
+            foreach (FileInfo file in files)
+            {
+                file.Delete();
+            }
+
+            if (recursive)
+            {
+                var directories = dir.GetDirectories();
+                foreach (DirectoryInfo di in directories)
+                {
+                    DeleteDirectory(di.FullName, true);
+                    di.Delete();
+                }
+            }
+        }
+
+        private void DeleteFiles(string path, params string[] extensions)
+        {
+            // https://stackoverflow.com/a/13301088
+            var ext = extensions.ToList();
+            var files = Directory.EnumerateFiles(path, "*.*", SearchOption.TopDirectoryOnly)
+                                 .Where(s => ext.Contains(Path.GetExtension(s).TrimStart().ToLowerInvariant()));
+
+            // https://stackoverflow.com/a/8132800
+            foreach (var file in files)
+            {
+                try
+                {
+                    File.SetAttributes(file, FileAttributes.Normal);
+                    File.Delete(file);
+                }
+                catch { }
+            }
+        }
     }
 }